implemented pdf generation

+++ controllers/admin.js
@@ -1,3 +1,5 @@
+const fileHelper = require('../util/file');
+
 const {validationResult} = require('express-validator/check');
 
 const Product = require('../models/product');
@@ -19,7 +21,7 @@ exports.postAddProduct = (req, res, next) => {
     const price = req.body.price;
     const description = req.body.description;
 
-    if (! image) {
+    if (!image) {
         return res
             .status(422)
             .render('admin/edit-product', {
@@ -145,6 +147,7 @@ exports.postEditProduct = (req, res, next) => {
             product.price = updatedPrice;
             product.description = updatedDesc;
             if (image) {
+                fileHelper.deleteFile(product.imageUrl);
                 product.imageUrl = image.path;
             }
 
@@ -182,7 +185,15 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({_id: prodId, userId: req.user._id})
+    Product
+        .findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('File not found!'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({_id: prodId, userId: req.user._id})
+        })
         .then(() => {
             console.log('DESTROYED PRODUCT');
             res.redirect('/admin/products');
